\name{print.APLE}
\alias{print.APLE}
\title{
	Printing the result of APLE
}
\description{
	Print method for APLE objects.
}
\usage{
	\method{print}{APLE}(x, ...)
}
\arguments{
  \item{x}{
	An APLE object produced by APLE() or VFSAPLE().
}
  \item{\dots}{
  	Other arguments to print
  }
}
\value{
	It returns the APLE x invisibly.
}

\author{
	Sarah Goslee
}

\seealso{
\code{\link{APLE}}, \code{\link{VFSAPLE}}, \code{\link{summary.APLE}},
}
\examples{
}
\keyword{ print }

