\name{summary.APLE}
\alias{summary.APLE}
\title{
	Summarize the result of APLE
}
\description{
	Summary method for APLE objects
}
\usage{
	\method{summary}{APLE}(object, ...)
}
\arguments{
  \item{object}{
	A APLE object produced by APLE().
	}
  \item{\dots}{
  	Other arguments to summary
  }
}

\details{
	Calculates means for phosphorus loss.
}

\value{
	Summary of APLE object.
}

\author{
	Sarah Goslee
}

\seealso{
\code{\link{APLE}}, \code{\link{VFSAPLE}}, \code{\link{print.APLE}},
}
\examples{

}

\keyword{ methods }

